put many examples under test
